# server/openapi.yaml
openapi: 3.1.0
info:
  title: Research Optimizer API
  description: |
    Public-facing API for the Research Optimizer platform. Served by the main
    server process which handles HTTPS termination, authentication, and proxying
    to internal services (gateway, cms).

    All `/api` endpoints require authentication unless noted otherwise. Authentication
    is via session cookie (after OAuth login) or `X-API-Key` header.
  version: 1.0.0
servers:
  - url: https://localhost
    description: Local development
paths:
...
  /api/agents:
    post:
      operationId: createAgent
      summary: Create an agent
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                promptID:
                  type: integer
                modelParameters:
                  type: object
                tools:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Agent created
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: getAgents
      summary: List agents
      description: User's agents plus global agents (userID=null).
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: List of agents
  /api/agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getAgent
      summary: Get agent by ID
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Agent details
        "404":
          description: Not found
    put:
      operationId: updateAgent
      summary: Update an agent
      description: Returns 403 for global agents.
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Agent updated
        "403":
          description: Cannot modify global agent
    delete:
      operationId: deleteAgent
      summary: Delete an agent
      description: Cascading delete.
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Agent deleted
  /api/conversations:
    post:
      operationId: createConversation
      summary: Create a conversation
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                agentID:
                  type: integer
      responses:
        "201":
          description: Conversation created
    get:
      operationId: getConversations
      summary: List conversations (paginated)
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated conversation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      limit:
                        type: integer
                      offset:
                        type: integer
  /api/conversations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getConversation
      summary: Get conversation by ID
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Conversation details
        "404":
          description: Not found
    put:
      operationId: updateConversation
      summary: Update a conversation
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        "200":
          description: Conversation updated
    delete:
      operationId: deleteConversation
      summary: Delete a conversation
      description: Soft delete.
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Conversation deleted
  /api/conversations/{id}/context:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getContext
      summary: Get conversation context
      description: Returns conversation with all messages and message-linked resources.
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Conversation context
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation:
                    type: object
                  messages:
                    type: array
                    items:
                      type: object
                  resources:
                    type: array
                    items:
                      type: object
        "404":
          description: Conversation not found
...
  /api/conversations/{conversationId}/messages:
    parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: integer
    post:
      operationId: appendConversationMessage
      summary: Add a message
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, content]
              properties:
                role:
                  type: string
                content:
                  type: object
                parentID:
                  type: integer
      responses:
        "201":
          description: Message created
    get:
      operationId: getMessages
      summary: Get messages in a conversation
      description: Ordered by createdAt ASC.
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: List of messages
  /api/messages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateMessage
      summary: Update a message
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Message updated
    delete:
      operationId: deleteMessage
      summary: Delete a message
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Message deleted
  /api/resources:
    post:
      operationId: storeConversationResource
      summary: Create a resource
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type, content]
              properties:
                name:
                  type: string
                type:
                  type: string
                content:
                  type: string
                agentID:
                  type: integer
                messageID:
                  type: integer
                s3Uri:
                  type: string
                metadata:
                  type: object
      responses:
        "201":
          description: Resource created
  /api/resources/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getResource
      summary: Get resource by ID
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Resource details
    delete:
      operationId: deleteConversationResource
      summary: Delete a resource
      description: Cascading delete — removes vectors.
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Resource deleted
  /api/agents/{agentId}/resources:
    parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getResourcesByAgent
      summary: Get resources for an agent
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: List of resources
  /api/conversations/{conversationId}/vectors:
    parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: integer
    post:
      operationId: storeConversationVectors
      summary: Add vectors to a conversation
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vectors]
              properties:
                vectors:
                  type: array
                  items:
                    type: object
                    required: [content]
                    properties:
                      content:
                        type: string
                      embedding:
                        type: array
                        items:
                          type: number
                      resourceID:
                        type: integer
                      toolID:
                        type: integer
                      order:
                        type: integer
      responses:
        "201":
          description: Vectors created
    get:
      operationId: getVectorsByConversation
      summary: Get vectors for a conversation
      tags: [Conversations]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: List of vectors

  # ===== ADMIN =====
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set after OAuth login
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key (rsk_... format) for programmatic access
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string